secrets-migration/pyproject.toml
[tool.poetry]
name = "secrets-migration"
version = "0.1.0"
description = "Projeto para migrar segredos entre contas"
authors = ["Kelder Passos <kelder.passos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
botocore = "^1.35.13"
boto3 = "^1.35.13"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
